feat: move endgame components to the new view folder
