+++ pages/endgame/index.tsx
@@ -1,10 +1,10 @@
-import EndgameContainer from '@ses/containers/Endgame/EndgameContainer';
 import { fetchAnalytics } from '@ses/containers/Finances/api/queries';
 import { getYearsRange } from '@ses/containers/Finances/utils/utils';
 import { DateTime } from 'luxon';
 import React from 'react';
 // eslint-disable-next-line camelcase
 import { SWRConfig, unstable_serialize } from 'swr';
+import EndgameView from '@/views/Endgame/EndgameView';
 import type { Analytic } from '@ses/core/models/interfaces/analytic';
 
 interface EndgamePageProps {
@@ -16,7 +16,7 @@ interface EndgamePageProps {
 
 const EndgamePage: React.FC<EndgamePageProps> = ({ budgetTransitionAnalytics, yearsRange, initialYear, fallback }) => (
   <SWRConfig value={{ fallback }}>
-    <EndgameContainer
+    <EndgameView
       budgetTransitionAnalytics={budgetTransitionAnalytics}
       yearsRange={yearsRange}
       initialYear={initialYear}
