--- src/views/Endgame/EndameView.tsx
@@ -1,3 +1,0 @@
-const EndgameView: React.FC = () => <div>endgame</div>;
-
-export default EndgameView;
