+++ src/views/Endgame/components/IntroductoryHeadline/IntroductoryHeadline.tsx
