+++ src/views/Endgame/components/KeyChangesSections/KeyChangesSections.tsx
