+++ src/views/Endgame/components/KeyChangesSections/Sections/OperationsSection.tsx
