+++ src/views/Endgame/components/KeyChangesSections/SectionsEnum.ts
