+++ src/views/Endgame/components/LatestUpdatesSection/LatestUpdatesSection.tsx
