+++ src/views/Endgame/components/SectionHeader/SectionHeader.tsx
