+++ src/views/Endgame/components/TransitionHeader/TransitionHeader.tsx
