+++ src/views/Endgame/types.ts
