+++ src/views/Endgame/utils/index.ts
